tokens/pyproject.toml
[project]
name = "tokens"
version = "0.1.0"
description = "tokens using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"

dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",

    # Optional: if you are using langchain-litellm, keep this:
    # "langchain-litellm>=0.3.0",
]

[project.scripts]
tokens = "tokens.main:run"
run_crew = "tokens.main:run"
train = "tokens.main:train"
replay = "tokens.main:replay"
test = "tokens.main:test"
run_with_trigger = "tokens.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
